fix: migrated authorize endpoint
