+++ package.json
@@ -55,7 +55,7 @@
     "@peculiar/x509": "^1.12.3",
     "@planetscale/database": "1.19.0",
     "arctic": "^2.3.2",
-    "authhero": "^0.31.0",
+    "authhero": "^0.32.0",
     "aws4fetch": "^1.0.20",
     "bcryptjs": "^2.4.3",
     "fast-xml-parser": "^4.5.1",
